fix: EXPOSED-107 Inaccurate UByte column type mapping (#1808)
* fix: EXPOSED-107 Incorrect mapping for UByte data type

Currently, when attempting to insert a UByte value outside of the range [0, 127],
Exposed truncates the value by calling value.toByte() before sending it to the
DB, causing overflow. The value is stored successfully as a negative number because
all databases (except MySQL and MariaDB) don't support unsigned types natively,
which means Exposed is actually mapping to 1-byte TINYINT, which accepts
the range [-128, 127].

Note: SQL Server is an exception as that DB represents TINYINT as holding the unsigned range [0, 255]; so SQL Server requires no changes.

Change the default mapping to the next higher-up integer data type SMALLINT (technically
a 2-byte storage type) and remove the truncation conversions so that an accurate
value is sent/received to/from the database. To ensure that the intended behavior
cannot be overriden using exec() directly, a check constraint is auto-applied to
the column when registered if the database is not MySQL/MariaDB/SQL Server.

Oracle TINYINT is an alias for NUMBER(38), so this has been overriden to a reduced
type NUMBER(4).

+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/vendors/Default.kt
@@ -20,8 +20,12 @@ abstract class DataTypeProvider {
     /** Numeric type for storing 1-byte integers. */
     open fun byteType(): String = "TINYINT"
 
-    /** Numeric type for storing 1-byte unsigned integers. */
-    open fun ubyteType(): String = "TINYINT"
+    /** Numeric type for storing 1-byte unsigned integers.
+     *
+     * **Note:** If the database being used is not MySQL, MariaDB, or SQL Server, this will represent the 2-byte
+     * integer type.
+     */
+    open fun ubyteType(): String = "SMALLINT"
 
     /** Numeric type for storing 2-byte integers. */
     open fun shortType(): String = "SMALLINT"
